flu_01.0001
Cambio del texto de inglés a español

+++ database/migrations/2019_03_26_114509_create_ivas_table.php
@@ -15,7 +15,7 @@ class CreateIvasTable extends Migration
     {
         Schema::create('ivas', function (Blueprint $table) {
             $table->increments('id');
-            $table->decimal('iva', 8, 2);
+            $table->integer('iva');
             $table->timestamps();
             $table->softDeletes();
         });
